init models from new feature
