+++ src/recipe/Baked.Recipe.Service.Application/Binding/Rest/AddTargetParameterConvention.cs
@@ -1,0 +1,15 @@
+﻿using Baked.Domain.Configuration;
+using Baked.RestApi.Model;
+
+namespace Baked.Binding.Rest;
+
+public class AddTargetParameterConvention(string parameterName)
+    : IDomainModelConvention<MethodModelContext>
+{
+    public void Apply(MethodModelContext context)
+    {
+        if (!context.Method.TryGetSingle<ActionModel>(out var action)) { return; }
+
+        action.Parameter[parameterName] = new ParameterModel(parameterName, context.Type.CSharpFriendlyFullName, ParameterModelFrom.Services);
+    }
+}
