--- src/recipe/Baked.Recipe.Service.Application/Binding/RestBinding/RestBindingExtensions.cs
@@ -1,10 +1,0 @@
-using Baked.Business;
-using Baked.Business.DomainAssemblies;
-
-namespace Baked;
-
-public static class RestBindingExtensions
-{
-    public static RestBindingFeature RestBinding(this BusinessConfigurator _) =>
-        new();
-}
