+++ src/recipe/Baked.Recipe.Service.Application/CodingStyle/ObjectAsJson/ObjectAsJsonCodingStyleFeature.cs
@@ -1,5 +1,4 @@
 ﻿using Baked.Architecture;
-using Baked.Business;
 using FluentNHibernate.Conventions.Helpers;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.OpenApi.Models;
