+++ src/recipe/Baked.Recipe.Service.Application/CodingStyle/RecordsAreDtos/RecordsAreDtosCodingStyleFeature.cs
@@ -1,5 +1,4 @@
 using Baked.Architecture;
-using Baked.Business;
 
 namespace Baked.CodingStyle.RecordsAreDtos;
