+++ src/recipe/Baked.Recipe.Service.Application/CodingStyle/UseBuiltInTypes/UseBuiltInTypesCodingStyleFeature.cs
@@ -1,5 +1,4 @@
 ﻿using Baked.Architecture;
-using Baked.Business;
 using FluentNHibernate.Conventions.Helpers;
 using Microsoft.Extensions.DependencyInjection;
 using Newtonsoft.Json.Converters;
