+++ src/recipe/Baked.Recipe.Service.Application/CodingStyle/UseNullableTypes/UseNullableTypesCodingStyleFeature.cs
@@ -1,5 +1,4 @@
 ﻿using Baked.Architecture;
-using Baked.Business;
 using System.Diagnostics.CodeAnalysis;
 using System.Reflection;
