+++ src/recipe/Baked.Recipe.Service.Application/Logging/Request/MappedMethodLogScopeMiddleware.cs
@@ -1,5 +1,4 @@
-﻿using Baked.Business;
-using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.Features;
 using Microsoft.Extensions.Logging;
