+++ src/recipe/Baked.Recipe.Service.Application/RestApi/Model/ActionModel.cs
@@ -3,15 +3,15 @@
 namespace Baked.RestApi.Model;
 
 [AttributeUsage(AttributeTargets.Method)]
-public class ActionModel(string method, string[] routeParts,
+public class ActionModel(string method, string[] routeParts, string findTargetStatement,
     string[]? additionalAttributes = default,
     string[]? preparationStatements = default
 ) : Attribute
 {
-    public ActionModel(string id, HttpMethod method, IEnumerable<string> routeParts, string returnType, bool returnIsAsync, bool returnIsVoid, IEnumerable<ParameterModel> parameters,
+    public ActionModel(string id, HttpMethod method, IEnumerable<string> routeParts, string findTargetStatement, string returnType, bool returnIsAsync, bool returnIsVoid, IEnumerable<ParameterModel> parameters,
         IEnumerable<string>? additionalAttributes = default,
         IEnumerable<string>? preparationStatements = default
-    ) : this(method.ToString(), [.. routeParts],
+    ) : this(method.ToString(), [.. routeParts], findTargetStatement,
         additionalAttributes: additionalAttributes?.ToArray(),
         preparationStatements: preparationStatements?.ToArray()
     )
@@ -27,10 +27,27 @@ public class ActionModel(string method, string[] routeParts,
     public List<string> RouteParts { get; set; } = [.. routeParts];
     public Func<string, string> RoutePartStylizer { get; set; } = s => s.Kebaberize();
     public string ReturnType { get; set; } = default!;
+
+    /// <summary>
+    /// Do NOT set this property directly from the attribute definition, e.g.,
+    /// `[ActionModel(..., ReturnIsAsync = true, ...)]`. Initial value is
+    /// always overridden by the value comes from reflection.
+    ///
+    /// Use conventions to set a custom value.
+    /// </summary>
     public bool ReturnIsAsync { get; set; } = default!;
+
+    /// <summary>
+    /// Do NOT set this property directly from the attribute definition, e.g.,
+    /// `[ActionModel(..., ReturnIsVoid = true, ...)]`. Initial value is always
+    /// overridden by the value comes from reflection.
+    ///
+    /// Use conventions to set a custom value.
+    /// </summary>
     public bool ReturnIsVoid { get; set; } = default!;
+
     public Func<string, string> ReturnResultRenderer { get; set; } = resultExpression => resultExpression;
-    public string FindTargetStatement { get; set; } = "target";
+    public string FindTargetStatement { get; set; } = findTargetStatement;
     public bool UseForm { get; set; } = false;
     public bool UseRequestClassForBody { get; set; } = true;
     public int Order { get; set; } = 0;
@@ -49,21 +66,13 @@ public class ActionModel(string method, string[] routeParts,
     public IEnumerable<ParameterModel> NonBodyParameters => NonServiceParameters.Where(p => p.From != ParameterModelFrom.BodyOrForm);
     public IEnumerable<ParameterModel> InvokedMethodParameters => Parameters.Where(p => p.IsInvokeMethodParameter);
 
-    internal ActionModel Init(string id, string defaultReturnType, bool defaultReturnIsAsync, bool defaultReturnIsVoid, IEnumerable<ParameterModel> parameters)
+    internal ActionModel Init(string id, string returnType, bool returnIsAsync, bool returnIsVoid, IEnumerable<ParameterModel> parameters)
     {
         Id = id;
         Name ??= id;
-        ReturnType ??= defaultReturnType;
-
-        // Below two has an exceptional case. It is possible for someone to set
-        // these values from an attribute definition, e.g., [ActionModel(...,
-        // ReturnIsAsync = true)], but this is not expected in practice. So
-        // initial values are always overridden by the default values. The only
-        // way to change them will be through conventions, because conventions
-        // run after `Init` is called.
-        ReturnIsAsync = defaultReturnIsAsync;
-        ReturnIsVoid = defaultReturnIsVoid;
-
+        ReturnType ??= returnType;
+        ReturnIsAsync = returnIsAsync;
+        ReturnIsVoid = returnIsVoid;
         Parameter ??= parameters.ToDictionary(p => p.Id);
 
         return this;
