+++ src/recipe/Baked.Recipe.Service.Application/RestApi/Model/ApiInputAttribute.cs
@@ -1,4 +1,4 @@
-﻿namespace Baked.Business;
+﻿namespace Baked.RestApi.Model;
 
 [AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct | AttributeTargets.Enum)]
 public class ApiInputAttribute : Attribute;
