+++ src/recipe/Baked.Recipe.Service/Binding/RequestResponseExampleData.cs
@@ -1,3 +1,3 @@
-﻿namespace Baked.Business;
+﻿namespace Baked.Binding;
 
 public record RequestResponseExampleData(string? Request, string? Response);
