+++ src/recipe/Baked.Recipe.Service/Binding/RequestResponseExamples.cs
@@ -1,3 +1,3 @@
-﻿namespace Baked.Business;
+﻿namespace Baked.Binding;
 
 public class RequestResponseExamples : Dictionary<string, RequestResponseExampleData>;
